+++ apps/web/src/components/tx-flow/flows/TokenTransfer/ReviewSpendingLimitTx.tsx
@@ -23,7 +23,7 @@ import useOnboard from '@/hooks/wallets/useOnboard'
 import { asError } from '@safe-global/utils/services/exceptions/utils'
 import TxCard from '@/components/tx-flow/common/TxCard'
 import { TxModalContext } from '@/components/tx-flow'
-import { type SubmitCallback } from '@/components/tx/SignOrExecuteForm/SignOrExecuteFormV2'
+import { type SubmitCallback } from '@/components/tx/SignOrExecuteForm/SignOrExecuteForm'
 import { TX_EVENTS, TX_TYPES } from '@/services/analytics/events/transactions'
 import { isWalletRejection } from '@/utils/wallets'
 import { safeParseUnits } from '@safe-global/utils/utils/formatters'
