+++ apps/web/src/components/tx/SignOrExecuteForm/SignForm.tsx
@@ -1,5 +1,5 @@
 import madProps from '@/utils/mad-props'
-import { type ReactElement, ReactNode, type SyntheticEvent, useContext, useMemo, useState } from 'react'
+import { type ReactElement, type ReactNode, type SyntheticEvent, useContext, useMemo, useState } from 'react'
 import { CircularProgress, Box, Button, Divider, Tooltip } from '@mui/material'
 import ErrorMessage from '@/components/tx/ErrorMessage'
 import { trackError, Errors } from '@/services/exceptions'
